LPD-63146 Move and rename
