+++ modules/apps/layout/layout-api/src/main/java/com/liferay/layout/importer/PortletPermissionsImporter.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.layout.importer.helper;
+package com.liferay.layout.importer;
 
 import java.util.List;
 import java.util.Map;
@@ -12,7 +12,7 @@ import java.util.Set;
 /**
  * @author Lourdes Fernández Besada
  */
-public interface PortletPermissionsImporterHelper {
+public interface PortletPermissionsImporter {
 
 	public void importPortletPermissions(
 			long plid, String portletId, Set<String> warningMessages,
